requirements.txt
Django==5.2
pillow==11.2.1
environs==14.1.1
django-taggit==6.1.0
django-nested-admin==4.1.1
